Refactored rest of packages, replaced built-in logger with my equivalent library
