+++ rs-timers-data.cfg
@@ -35,7 +35,7 @@ timers {
 		name:Wicked Hood
 		audio:false
 		tab:0
-		latestreset:1503100800000
+		latestreset:1503187200000
 	}
 	Daily {
 		name:Vis Wax
