--- src/com/minepop/talkar/util/logging/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Talkarcabbage
- *
- */
-package com.minepop.talkar.util.logging;
