--- src/com/minepop/talkar/util/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Talkarcabbage
- * This package contains the majority of transferable utility classes I write.
- */
-package com.minepop.talkar.util;
